Complete RESTful service

+++ app_socialnet/backend/controllers/feed.js
@@ -3,6 +3,7 @@ const fs = require("fs");
 const path = require("path");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -44,22 +45,28 @@ exports.postPost = (req, res, next) => {
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
 
   const post = new Post({
     title,
     content,
     imageUrl,
-    creator: {
-      name: "hackfeed",
-    },
+    creator: req.userId,
   });
 
   post
     .save()
-    .then((result) => {
+    .then(() => User.findById(req.userId))
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save;
+    })
+    .then(() => {
       res.status(201).json({
         message: "Post created successfully",
-        post: result,
+        post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -120,6 +127,12 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -152,9 +165,21 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
 
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post.imageUrl);
 
       return Post.findByIdAndRemove(postId);
+    })
+    .then(() => User.findById(req.userId))
+    .then((user) => {
+      user.posts.pull(postId);
+
+      return user.save();
     })
     .then(() => res.status(200).json({ message: "Deleted post." }))
     .catch((err) => {
